django_framework/django_rest_framework_pro/requirements.txt
django==2.1.5
django-extensions
django-filter
django-redis
django-rest-framework-mongoengine
djangorestframework

djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0
pygments



